AWS: add HadoopConfigurableGlueCatalog
